+++ src/pages/LetterReceptionPage/ReplyToLetter/index.tsx
@@ -1,10 +1,8 @@
-import { useForm, FormProvider } from 'react-hook-form';
+import { useFormContext } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
 import { toast } from 'react-toastify';
 import { useEffect } from 'react';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { zodResolver } from '@hookform/resolvers/zod';
-import z from 'zod';
 import { isAxiosError } from 'axios';
 import LetterCard from '@/components/LetterCard';
 import Navbar from '@/components/Navbar';
@@ -14,39 +12,17 @@ import LetterLengthDate from '@/components/LetterLengthDate';
 import LetterHeader from '@/components/LetterHeader';
 import letterAPI from '@/api/letter/apis';
 import LoadingSpinner from '@/components/LoadingSpinner';
-import { letterWrite } from '@/constants/schemaLiteral';
 import letterOptions from '@/api/letter/queryOptions';
 import { ROUTER_PATHS } from '@/router';
 import ERROR_RESPONSES from '@/constants/errorMessages';
 import useBoolean from '@/hooks/useBoolean';
 import BottomSheet from '@/components/BottomSheet';
-import LetterContent from '../components/LetterContent';
+import { ReplyInputs } from '..';
 import useLetterWithTags from '../hooks/useLetterWithTags';
-import ReplyImage from './ReplyImage';
-import style from './styles';
+import LetterContent from '../components/LetterContent';
 import ReceivedAccordionLetter from './ReceivedAccordionLetter';
-
-const L = letterWrite;
-
-const replySchema = z.object({
-  replyContent: z
-    .string()
-    .min(L.content.min.value, { message: L.content.min.message })
-    .max(L.content.max.value, { message: L.content.max.message }),
-  image: z
-    .any()
-    .optional()
-    .refine(
-      (files) => !files || files[0].size <= L.image.maxFileSize.value,
-      L.image.maxFileSize.message,
-    )
-    .refine(
-      (files) => !files || L.image.acceptType.list.includes(files[0].type),
-      L.image.acceptType.message,
-    ),
-});
-
-export type ReplyInputs = z.infer<typeof replySchema>;
+import style from './styles';
+import ReplyImage from './ReplyImage';
 
 interface ReplyToLetterProps {
   letterId: number;
@@ -60,19 +36,12 @@ const ReplyToLetter = ({ letterId, onPrev }: ReplyToLetterProps) => {
   const { value, on, off } = useBoolean(false);
   const queryClient = useQueryClient();
 
-  const methods = useForm<ReplyInputs>({
-    resolver: zodResolver(replySchema),
-    defaultValues: {
-      replyContent: '',
-    },
-  });
-
   const {
     handleSubmit,
     register,
     watch,
     formState: { errors },
-  } = methods;
+  } = useFormContext<ReplyInputs>();
 
   const { mutateAsync: patchReply, isPending } = useMutation({
     mutationFn: letterAPI.patchReceptionReply,
@@ -132,71 +101,64 @@ const ReplyToLetter = ({ letterId, onPrev }: ReplyToLetterProps) => {
   }, [errors]);
 
   return (
-    <FormProvider {...methods}>
-      <LetterContent isBlock={true}>
-        <form onSubmit={handleSubmit(on)}>
-          <div css={style.letter}>
-            <ReceivedAccordionLetter receptionLetter={receptionLetter} />
-            <LetterCard isOpen={true} css={style.card}>
-              <LetterHeader
-                title="To"
-                nickname={receptionLetter.senderNickname}
-              />
-              <LetterTextarea
-                {...register('replyContent')}
-                name="replyContent"
-                placeholder="하고싶은 이야기를 답장으로 적어보세요. (10자 이상)"
-              />
-              <LetterLengthDate letterLength={watch('replyContent').length} />
-              <LetterHeader
-                title="From"
-                titlePosition="right"
-                nickname={receptionLetter.receiverNickname}
-              />
-              <ReplyImage />
-            </LetterCard>
-          </div>
-          <Navbar css={style.navbar}>
-            <Button
-              type="button"
-              variant="secondary"
-              size="sm"
-              onClick={onPrev}
-            >
-              취소
-            </Button>
-            <Button
-              disabled={isPending}
-              type="submit"
-              variant="primary"
-              size="sm"
-            >
-              {isPending ? <LoadingSpinner /> : '답장 보내기'}
-            </Button>
-          </Navbar>
-        </form>
-        <BottomSheet open={value} onOpen={on} onClose={off}>
-          <BottomSheet.Title>답장을 보낼까요?</BottomSheet.Title>
-          <BottomSheet.Description>
-            낯선이에게 답장은 한번만 가능해요
-          </BottomSheet.Description>
-          <BottomSheet.ButtonSection>
-            <Button variant="cancel" onClick={off}>
-              취소
-            </Button>
-            <Button
-              variant="primary"
-              onClick={() => {
-                handleSubmit(onSubmit)();
-                off();
-              }}
-            >
-              보내기
-            </Button>
-          </BottomSheet.ButtonSection>
-        </BottomSheet>
-      </LetterContent>
-    </FormProvider>
+    <LetterContent isBlock={true}>
+      <form onSubmit={handleSubmit(on)}>
+        <div css={style.letter}>
+          <ReceivedAccordionLetter receptionLetter={receptionLetter} />
+          <LetterCard isOpen={true} css={style.card}>
+            <LetterHeader
+              title="To"
+              nickname={receptionLetter.senderNickname}
+            />
+            <LetterTextarea
+              {...register('replyContent')}
+              name="replyContent"
+              placeholder="하고싶은 이야기를 답장으로 적어보세요. (10자 이상)"
+            />
+            <LetterLengthDate letterLength={watch('replyContent').length} />
+            <LetterHeader
+              title="From"
+              titlePosition="right"
+              nickname={receptionLetter.receiverNickname}
+            />
+            <ReplyImage />
+          </LetterCard>
+        </div>
+        <Navbar css={style.navbar}>
+          <Button type="button" variant="secondary" size="sm" onClick={onPrev}>
+            취소
+          </Button>
+          <Button
+            disabled={isPending}
+            type="submit"
+            variant="primary"
+            size="sm"
+          >
+            {isPending ? <LoadingSpinner /> : '답장 보내기'}
+          </Button>
+        </Navbar>
+      </form>
+      <BottomSheet open={value} onOpen={on} onClose={off}>
+        <BottomSheet.Title>답장을 보낼까요?</BottomSheet.Title>
+        <BottomSheet.Description>
+          낯선이에게 답장은 한번만 가능해요
+        </BottomSheet.Description>
+        <BottomSheet.ButtonSection>
+          <Button variant="cancel" onClick={off}>
+            취소
+          </Button>
+          <Button
+            variant="primary"
+            onClick={() => {
+              handleSubmit(onSubmit)();
+              off();
+            }}
+          >
+            보내기
+          </Button>
+        </BottomSheet.ButtonSection>
+      </BottomSheet>
+    </LetterContent>
   );
 };
